pyproject.toml
[tool.poetry]
name = "deeprootgen"
version = "0.1.0"
description = "A simulation model for the digital reconstruction of 3D root system architectures. Integrated with a simulation-based inference generative deep learning model."
authors = ["James Bristow <james.bristow@plantandfood.co.nz>"]
license = "GPL-3.0-only"
readme = "README.md"
repository = "https://github.com/JBris/deep-root-gen"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
kaleido = "0.2.1"
scikit-learn = "^1.5.1"
numpy = "^1.26.4"
pandas = "^2.2.2"
networkx = "^3.3"
dash = "^2.17.1"
plotly = "^5.23.0"
dash-core-components = "^2.0.0"
dash-html-components = "^2.0.0"
pillow = "^10.4.0"
scipy = "<=1.11.4"
joblib = "^1.4.2"
sqlalchemy = "^2.0.32"
hydra-core = "^1.3.2"
mlflow = "^2.15.1"
alembic = "^1.13.2"
celery = "^5.4.0"
boto3 = "^1.34.162"
botocore = "^1.34.162"
omegaconf = "^2.3.0"
seaborn = "<0.13"
graphviz = "^0.20.3"
prometheus-flask-exporter = "^0.23.1"
prometheus-client = "^0.20.0"
pyyaml = "^6.0.2"
prefect = "^2.20.2"
httpx = "^0.27.0"
pydantic = "^2.8.2"
dask = "^2024.8.0"
prefect-dask = "^0.2.10"
distributed = "^2024.8.0"
optuna = "^3.6.1"
bentoml = "^1.3.2"
psycopg2-binary = "^2.9.9"
arviz = "^0.19.0"
pymc = "^5.16.2"
pytensor = "^2.25.2"
dash-bootstrap-components = "^1.6.0"
dash-bootstrap-templates = "^1.2.3"


[tool.poetry.group.torch.dependencies]
sbi = "^0.22.0"
torch = "^2.4.0"
pyro-ppl = "^1.9.1"
pyro-api = "^0.1.2"
nflows = "^0.14"
tensorboard = "^2.17.1"
torch-geometric = "^2.5.3"


[tool.poetry.group.dev.dependencies]
ipython = "^8.26.0"
ipywidgets = "^8.1.3"
jupyter-core = "^5.7.2"
black = "^24.8.0"
pytest = "^8.3.2"
flake8 = "^7.1.1"
coverage = "^7.6.1"
pre-commit = "^3.8.0"
sphinx-rtd-theme = "^2.0.0"
pytest-cov = "^5.0.0"
sphinx = "<8"
autoflake = "^2.3.1"
mypy = "^1.11.1"
sphinx-argparse = "^0.5.2"
sphinxcontrib-applehelp = "^2.0.0"
sphinxcontrib-devhelp = "^2.0.0"
sphinxcontrib-htmlhelp = "^2.1.0"
sphinxcontrib-jsmath = "^1.0.1"
sphinxcontrib-qthelp = "^2.0.0"
sphinxcontrib-serializinghtml = "^2.0.0"
pytest-mock = "^3.14.0"
pytest-xdist = "^3.6.1"
pytest-env = "^1.1.3"
types-pyyaml = "^6.0.12.20240808"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
